backend/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
aiofiles=24.1.0=pyhd8ed1ab_1
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.4=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pyhd8ed1ab_1
anyio=4.9.0=pyh29332c3_0
attrs=25.3.0=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2025.4.26=h4c7d964_0
cachecontrol=0.14.3=pypi_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=pyhd8ed1ab_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
cryptography=45.0.3=pypi_0
dataclasses-json=0.6.7=pypi_0
deprecation=2.1.0=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pyhff2d567_1
ecdsa=0.19.1=pypi_0
email-validator=2.2.0=pyhd8ed1ab_1
email_validator=2.2.0=hd8ed1ab_1
et-xmlfile=2.0.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
fastapi=0.115.12=pyh29332c3_0
fastapi-cli=0.0.7=pyhd8ed1ab_0
filelock=3.18.0=pypi_0
firebase-admin=6.8.0=pypi_0
frozenlist=1.6.0=pypi_0
google-api-core=2.25.0rc1=pypi_0
google-api-python-client=2.170.0=pypi_0
google-auth=2.40.2=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-cloud-core=2.4.3=pypi_0
google-cloud-firestore=2.20.2=pypi_0
google-cloud-storage=3.1.0=pypi_0
google-crc32c=1.7.1=pypi_0
google-resumable-media=2.7.2=pypi_0
googleapis-common-protos=1.70.0=pypi_0
gotrue=2.12.0=pypi_0
greenlet=3.2.2=pypi_0
groq=0.26.0=pypi_0
grpcio=1.71.0=pypi_0
grpcio-status=1.71.0=pypi_0
h11=0.16.0=pyhd8ed1ab_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httplib2=0.22.0=pypi_0
httptools=0.6.4=py311he736701_0
httpx=0.28.1=pyhd8ed1ab_0
httpx-sse=0.4.0=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
iniconfig=2.1.0=pypi_0
jinja2=3.1.6=pyhd8ed1ab_0
jiter=0.10.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=0.3.25=pypi_0
langchain-community=0.3.24=pypi_0
langchain-core=0.3.63=pypi_0
langchain-groq=0.3.2=pypi_0
langchain-text-splitters=0.3.8=pypi_0
langsmith=0.3.43=pypi_0
libexpat=2.7.0=he0c23c2_0
libffi=3.4.6=h537db12_1
liblzma=5.8.1=h2466b09_1
libsqlite=3.50.0=h67fdade_0
libzlib=1.3.1=h2466b09_2
lxml=5.4.0=pypi_0
markdown-it-py=3.0.0=pyhd8ed1ab_1
markupsafe=3.0.2=py311h5082efb_1
marshmallow=3.26.1=pypi_0
mdurl=0.1.2=pyhd8ed1ab_1
msgpack=1.1.0=pypi_0
multidict=6.4.4=pypi_0
mypy-extensions=1.1.0=pypi_0
numpy=2.2.6=pypi_0
openai=1.82.1=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.5.0=ha4e3fda_1
orjson=3.10.18=pypi_0
packaging=24.2=pypi_0
pandas=2.3.0=pypi_0
pdfkit=1.0.0=pypi_0
pillow=11.2.1=pypi_0
pip=25.1.1=pyh8b19718_0
pluggy=1.6.0=pypi_0
postgrest=1.0.2=pypi_0
propcache=0.3.1=pypi_0
proto-plus=1.26.1=pypi_0
protobuf=5.29.5=pypi_0
psycopg=3.2.9=pypi_0
psycopg-binary=3.2.9=pypi_0
pyarrow=20.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.4=pyh3cfb1c2_0
pydantic-core=2.33.2=py311hc4022dc_0
pydantic-settings=2.9.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyjwt=2.10.1=pypi_0
pymupdf=1.26.0=pypi_0
pyparsing=3.2.3=pypi_0
pytest=8.3.5=pypi_0
pytest-mock=3.14.1=pypi_0
python=3.11.12=h3f84c4b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
python-dotenv=1.1.0=pyh29332c3_1
python-jose=3.5.0=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.20=pyhff2d567_0
python_abi=3.11=7_cp311
pytz=2025.2=pypi_0
pyyaml=6.0.2=py311h5082efb_2
realtime=2.4.3=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=14.0.0=pyh29332c3_0
rich-toolkit=0.11.3=pyh29332c3_0
rsa=4.9.1=pypi_0
setuptools=80.8.0=pyhff2d567_0
shellingham=1.5.4=pyhd8ed1ab_1
six=1.17.0=pypi_0
sniffio=1.3.1=pyhd8ed1ab_1
sqlalchemy=2.0.41=pypi_0
starlette=0.46.2=pyh81abbef_0
storage3=0.11.3=pypi_0
strenum=0.4.15=pypi_0
supabase=2.15.2=pypi_0
supafunc=0.9.4=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.1.2=pypi_0
tk=8.6.13=h2c6b04d_2
together=1.5.8=pypi_0
tqdm=4.67.1=pypi_0
typer=0.15.4=pypi_0
typer-slim=0.16.0=pyhe01879c_0
typer-slim-standard=0.16.0=hf964461_0
typing-extensions=4.13.2=h0e9735f_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.1=pyhd8ed1ab_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
uritemplate=4.1.1=pypi_0
urllib3=2.4.0=pypi_0
uvicorn=0.34.2=pyh5737063_0
uvicorn-standard=0.34.2=h5737063_0
vc=14.3=h2b53caa_26
vc14_runtime=14.42.34438=hfd919c2_26
vs2015_runtime=14.42.34438=h7142326_26
watchfiles=1.0.5=py311h533ab2d_0
websockets=14.2=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
whisper=1.1.10=pypi_0
yaml=0.2.5=h8ffe710_2
yarl=1.20.0=pypi_0
zstandard=0.23.0=pypi_0
